Fixed minor errors in PackedArray/PackedFileArray Added call to checkInvariant in functions PackedArray:save/load, to chck if data-structure is consistent. PackedFileArray::get make chaeck for index-out-of-range in all configurations Added function PackedFileArray::checkInvariant (essential theame as PackedArray) and called in constructor Added unittest for PackedFileArray!!

+++ Test/UnitTest/TestBitSet/TestBitSet.cpp
@@ -676,7 +676,7 @@ namespace TestBitSet {
       {
         for(double capacity = 10; capacity < 650000; capacity *= 1.4) {
           OUTPUT(_T("testAllFileBitSetIndices(%.0lf"), capacity);
-          testAllFileBitSetIndices(capacity);
+          testAllFileBitSetIndices((UINT)capacity);
         }
         return;
       }
